chore: remove models from deprecated code
