--- aws_lambda_powertools/helper/__init__.py
@@ -1,2 +1,0 @@
-"""Collection of reusable code shared across powertools utilities
-"""
