+++ aws_lambda_powertools/logging/__init__.py
@@ -1,6 +1,5 @@
 """Logging utility
 """
-from ..helper.models import MetricUnit
 from .logger import Logger
 
-__all__ = ["MetricUnit", "Logger"]
+__all__ = ["Logger"]
